Add files via upload
The train Ticket Reservation System is a database system project that aims to provide an 
efficient and user-friendly solution for managing train tickets, bookings, and passenger 
information. Train Ticket Reservation is the system where users can reserve railway seats. 
This is the web browser where they can reserve it online. Designed a train ticket booking, 
reservation, and cancellation system for users who wish to travel via the Railways. It is 
written in PHP and MYSQL, and for styling HTML and CSS, JAVASCRIPT.